Calculate ranges x cells tile overlap matrix

Calculate ranges x cells tile overlap matrix

tileMatrix(
  fragments,
  ranges,
  zero_based_coords = TRUE,
  explicit_tile_names = FALSE
)

Arguments

fragments

Input fragments object

ranges

GRanges object with the ranges to overlap including a metadata column tile_width, or a list/data frame with columns chr, start, end, and tile_width. Must be non-overlapping and sorted by (chr, start), with chromosomes ordered according to the chromosome names of fragments

zero_based_coords

Boolean for whether the input ranges are in a 0-based or a 1-based coordinate system. (1-based will be converted to 0-based) (see http://genome.ucsc.edu/blog/the-ucsc-genome-browser-coordinate-counting-systems/)

explicit_tile_names

Boolean for whether to add rownames to the output matrix in format e.g chr1:500-1000, where start and end coords are given in a 0-based coordinate system. For whole-genome Tile matrices the names will take ~5 seconds to generate and take up 400MB of memory. Note that either way, tile names will be written when the matrix is saved.

Value

Iterable matrix object with dimension ranges x cells. When saved, the column names will be in the format chr1:500-1000, where start and end coords are given in a 0-based coordinate system.

Note

When calculating the matrix directly from a fragments tsv, it's necessary to first call select_chromosomes in order to provide the ordering of chromosomes to expect while reading the tsv.