Calculate gene-tile distances for ArchR gene activities

ArchR-style gene activity scores are based on a weighted sum of each tile according to the signed distance from the tile to a gene body. This function calculates the signed distances according to ArchR's default parameters.

geneScoreTilesArchR(genes, tile_size = 500, addArchRBug = FALSE)

Arguments

genes

GRanges object with gene start, end, and strand

tile_size

Size of tiles to consider

addArchRBug

Replicate ArchR bug in handling nested genes

Value

GRanges object with one range per tile, with additional metadata columns gene_idx (row index of the gene this tile corresponds to) and distance.

Distance is a signed distance calculated such that if the tile has a smaller start coordinate than the gene and the gene is on the + strand, distance will be negative and calculated as min(0, end(tile) - start(gene))

Details

ArchR's tile distance algorithm works as follows

Genes are extended 5kb upstream

Genes are linked to any tiles 1kb-100kb upstream + downstream, but tiles beyond a neighboring gene are not considered

Note: assumes the 1-based, end inclusive coordinate convention used by GRanges, so 500bp tiles run from bases 1-500, 501-1000, etc.