Calculate ArchR-compatible per-cell QC statistics

Calculate ArchR-compatible per-cell QC statistics

qc_scATAC(fragments, genes, blacklist)

Arguments

fragments

IterableFragments object

genes

GenomicRanges object or list/data.frame with columns chr, start, end

blacklist

GenomicRanges object or list/data.frame with columns chr, start, end

Value

data.frame with QC data

Details

This implementation mimics ArchR's default parameters. For uses requiring more flexibility to tweak default parameters, the best option is to re-implement this function with required changes. Output columns of data.frame:

cellName: cell name for each cell

nFrags: number of fragments per cell

subNucleosomal, monoNucleosomal, multiNucleosomal: number of fragments of size 1-146bp, 147-254bp, and 255bp + respectively. equivalent to ArchR's nMonoFrags, nDiFrags, nMultiFrags respectively

TSSEnrichment: AvgInsertInTSS / max(AvgInsertFlankingTSS, 0.1), where AvgInsertInTSS ReadsInTSS / 101 (window size), and AvgInsertFlankingTSS is ReadsFlankingTSS / (100*2) (window size). The max(0.1) ensures that very low-read cells do not get assigned spuriously high TSSEnrichment.

ReadsInPromoter: Number of reads from 2000bp upstream of TSS to 101bp downstream of TSS

ReadsInBlacklist: Number of reads in the provided blacklist region

ReadsInTSS: Number of reads overlapping the 101bp centered around each TSS

ReadsFlankingTSS: Number of reads overlapping 1901-2000bp +/- each TSS

Differences from ArchR: Note that ArchR by default uses a different set of annotations to derive TSS sites and promoter sites. This function uses just one annotation for gene start+end sites, so must be called twice to exactly re-calculate the ArchR QC stats.

ArchR's PromoterRatio and BlacklistRatio are not included in the output, as they can be easily calculated from ReadsInPromoter / nFrags and ReadsInBlacklist / nFrags. Similarly, ArchR's NucleosomeRatio can be calculated as (monoNucleosomal + multiNucleosomal) / subNucleosomal.